refactor: card title uses generic event
