+++ apps/journeys-admin/src/components/VisitorInfo/JourneyList/EventsCard/GenericEvent/GenericEvent.tsx
@@ -6,6 +6,7 @@ import TimelineContent from '@mui/lab/TimelineContent'
 import TimelineOppositeContent from '@mui/lab/TimelineOppositeContent'
 import TimelineDot from '@mui/lab/TimelineDot'
 import Typography from '@mui/material/Typography'
+import Box from '@mui/material/Box'
 import Stack from '@mui/material/Stack'
 import { format, parseISO } from 'date-fns'
 
@@ -13,7 +14,7 @@ interface Props {
   createdAt?: string
   label?: string | null
   value?: string | ReactNode | null
-  icon: ReactElement
+  icon?: ReactElement
   activity?: string
   duration?: string
   position?: 'start' | 'end'
@@ -47,37 +48,41 @@ export function GenericEvent({
         </Typography>
       </TimelineOppositeContent>
 
-      <TimelineSeparator
-        sx={{ justifyContent: position === 'end' ? 'flex-end' : 'flex-start' }}
-      >
-        <TimelineConnector
+      {icon != null ? (
+        <TimelineSeparator
           sx={{
-            display: position === 'end' ? 'none' : 'flex',
-            flexGrow: position === 'start' ? '0.5' : '1'
-          }}
-        />
-        <TimelineDot
-          sx={{
-            color: 'text.primary',
-            backgroundColor: 'transparent',
-            boxShadow: 'none',
-            m: 0
+            justifyContent: position === 'end' ? 'flex-end' : 'flex-start'
           }}
         >
-          {icon}
-        </TimelineDot>
-        <TimelineConnector
-          sx={{
-            display: position === 'start' ? 'none' : 'flex',
-            flexGrow: position === 'end' ? '0.5' : '1'
-          }}
-        />
-      </TimelineSeparator>
+          <TimelineConnector
+            sx={{
+              display: position === 'end' ? 'none' : 'flex'
+            }}
+          />
+          <TimelineDot
+            sx={{
+              color: 'text.primary',
+              backgroundColor: 'transparent',
+              boxShadow: 'none',
+              m: 0
+            }}
+          >
+            {icon}
+          </TimelineDot>
+          <TimelineConnector
+            sx={{
+              flexGrow: position === 'end' ? '0.5' : '1'
+            }}
+          />
+        </TimelineSeparator>
+      ) : (
+        <Box sx={{ width: '36px' }} />
+      )}
 
       <TimelineContent
-        sx={{ p: 2, display: 'flex', alignItems: 'center', py: 4 }}
+        sx={{ px: 2, py: 4, display: 'flex', alignItems: 'center' }}
       >
-        <Stack direction="column">
+        <Stack direction="column" sx={{ width: '100%' }}>
           <Stack direction="row">
             {activity != null && (
               <Typography variant="body2" sx={{ fontWeight: 'bold' }}>
