+++ apps/journeys-admin/src/components/VisitorInfo/JourneyList/EventsCard/GenericEvent/index.ts
