+++ apps/journeys-admin/src/components/VisitorInfo/JourneyList/EventsCard/TimelineEvent/TimelineEvent.tsx
@@ -13,7 +13,7 @@ import {
   messagePlatformToLabel,
   TimelineItem
 } from '../../utils'
-import { GenericEvent } from './GenericEvent'
+import { GenericEvent } from '../GenericEvent'
 
 interface Props {
   timelineItem: TimelineItem
